+++ ChangeLog.md
@@ -3,6 +3,7 @@ Unreleased changes
 * Add attachment detection to RightScript scaffold command
 * Check for updates with an OS and architecture specific YAML file so that MacOS
   builds taking longer do not result in update errors
+* Update command instructions now include sudo if it is necessary
 
 v1.4.5 / 2016-10-18
 -------------------